+++ python/src/lib.rs
@@ -12,11 +12,13 @@ mod io;
 mod memory_store;
 mod model;
 mod sled_store;
+mod sparql;
 mod store_utils;
 
 use crate::memory_store::*;
 use crate::model::*;
 use crate::sled_store::*;
+use crate::sparql::*;
 use pyo3::prelude::*;
 
 /// Oxigraph Python bindings
@@ -34,5 +36,9 @@ fn pyoxigraph(_py: Python<'_>, module: &PyModule) -> PyResult<()> {
     module.add_class::<PyQuad>()?;
     module.add_class::<PyMemoryStore>()?;
     module.add_class::<PySledStore>()?;
+    module.add_class::<PyVariable>()?;
+    module.add_class::<PyQuerySolutions>()?;
+    module.add_class::<PyQuerySolution>()?;
+    module.add_class::<PyQueryTriples>()?;
     io::add_to_module(module)
 }